[Android][Extension] Move extension tests to core_internal.
ExtensionEcho(renamed as ExtensionEchoInternal) uses core_internal interface
XWalkExtensionAndroid, so this extension along with the test should move
from core into core_internal.

Same thing ExtensionBroadcast.
