+++ test/android/core_internal/javatests/src/org/xwalk/core/internal/xwview/test/ExtensionBroadcastInternal.java
@@ -2,13 +2,13 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-package org.xwalk.core.xwview.test;
+package org.xwalk.core.internal.xwview.test;
 
 import org.xwalk.core.internal.extensions.XWalkExtensionAndroid;
 
-public class ExtensionBroadcast extends XWalkExtensionAndroid {
+public class ExtensionBroadcastInternal extends XWalkExtensionAndroid {
 
-    public ExtensionBroadcast() {
+    public ExtensionBroadcastInternal() {
         super("broadcast",
               "exports.setHandler = function(handler) {"
               + "  extension.setMessageListener(handler);"
